Fixed like bug
