+++ src/handleLikes.js
@@ -1,4 +1,4 @@
-const likesUrl = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/tG7RtMwoebPaqmjha3lr/likes/';
+const likesUrl = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/OqoSKrCAwmdPUD6ymYnk/likes/';
 
 const likesArray = [];
 const getLikes = async () => {
@@ -11,7 +11,7 @@ const getLikes = async () => {
 };
 
 const updateLikes = async (id, likeCounter) => {
-  await getLikes();
+  await getLikes(id);
   let counter = 0;
   likesArray.forEach((like) => {
     if (like.item_id === id) {
@@ -33,13 +33,15 @@ const sendData = async (id) => {
   });
 };
 
-document.addEventListener('click', (e) => {
+document.addEventListener('click', async (e) => {
   const likeID = Number(e.target.id);
-  if (e.target.classList.contains('fa-heart')) {
+  if (e.target.classList.contains('far')) {
     e.target.classList.remove('far');
     e.target.classList.add('fas');
-    updateLikes(likeID);
-    sendData(likeID);
+    const numLikes = e.target.parentElement.children[1].innerText;
+    e.target.parentElement.children[1].innerText = `${Number(numLikes) + 1}`;
+    await sendData(likeID);
+    await updateLikes(likeID);
   }
 });
